Cambiar double a long double y otros cambios de clang-tidy

+++ src/milk_type.cpp
@@ -3,14 +3,14 @@
 #include <cstdio>
 
 
-MilkType::MilkType(char typeId, long quota, double profit){
+MilkType::MilkType(char typeId, long quota, long double profit){
     this->typeId = typeId;
     this->quota = quota;
     this->profit = profit;
 }
 
 void MilkType::print(bool newline) const{
-    printf("<MilkType. typeId: %c, quota: %ld, profit: %lf>%s", typeId, quota, profit, newline?"\n":"");
+    printf("<MilkType. typeId: %c, quota: %ld, profit: %Lf>%s", typeId, quota, profit, newline?"\n":"");
 }
 
 char MilkType::getId() const{
@@ -19,6 +19,6 @@ char MilkType::getId() const{
 long MilkType::getMilkQuota() const{
     return quota;
 }
-double MilkType::getMilkProfit() const{
+long double MilkType::getMilkProfit() const{
     return profit;
 }
